# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ConsumerApp API",
        "version": "1.0"
    },
    "paths": {
...
        "/api/complex": {
            "get": {
                "tags": [
                    "Client"
                ],
                "description": "Получение списка ЖК",
                "responses": {
                    "200": {
                        "description": "Список ЖК найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "description": "Массив найденных ЖК",
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/ResidentialComplex"
                                                    }
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ResidentialComplex"
                ],
                "description": "Добавить ЖК",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResidentialComplexRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ЖК добавлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/ResidentialComplex"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/complex/{complex}": {
            "get": {
                "tags": [
                    "ResidentialComplex"
                ],
                "description": "Получение ЖК",
                "responses": {
                    "200": {
                        "description": "ЖК найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/ResidentialComplex"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ResidentialComplex"
                ],
                "description": "Отредактировать ЖК",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ResidentialComplexRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ЖК отредактирован",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/ResidentialComplex"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ResidentialComplex"
                ],
                "description": "Получение списка ЖК",
                "responses": {
                    "200": {
                        "description": "ЖК удалён",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "data": {
                                                    "type": "null"
                                                },
                                                "meta": {
                                                    "description": "Мета-теги",
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "ResidentialComplexRequest": {
                "properties": {
                    "title": {
                        "description": "Название ЖК",
                        "type": "string"
                    },
                    "address": {
                        "description": "Адрес ЖК",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "Flat": {
                "properties": {
                    "id": {
                        "description": "Идентификатор квартиры",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название квартиры",
                        "type": "string"
                    },
                    "status": {
                        "$ref": "#/components/schemas/FlatStatus"
                    },
                    "full_space": {
                        "description": "Вся площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "floor_count": {
                        "description": "Кол-во этажей",
                        "type": "integer"
                    },
                    "living_space": {
                        "description": "Жилая площадь",
                        "type": "number",
                        "format": "float"
                    },
                    "room_count": {
                        "description": "Кол-во комнат",
                        "type": "integer"
                    },
                    "balconyless_space": {
                        "description": "Площадьь без балконов и т.п.",
                        "type": "number",
                        "format": "float"
                    },
                    "residential_complex_id": {
                        "description": "Идентификатор ЖК, в котором располагается квартира",
                        "type": "integer"
                    },
                    "cost": {
                        "description": "Начальная цена квартиры",
                        "type": "integer"
                    },
                    "is_ready": {
                        "description": "Готова ли квартира",
                        "type": "boolean"
                    }
                },
                "type": "object"
            },
            "FlatStatus": {
                "properties": {
                    "id": {
                        "description": "Идентификатор статуса квартиры",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название статуса квартиры",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ResidentialComplex": {
                "properties": {
                    "id": {
                        "description": "Идентификатор ЖК",
                        "type": "integer"
                    },
                    "title": {
                        "description": "Название ЖК",
                        "type": "string"
                    },
                    "address": {
                        "description": "Адрес ЖК",
                        "type": "string"
                    },
                    "flats": {
                        "description": "Квартиры ЖК",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Flat"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}